+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## v0.0.3
+
+## Added
+
+- Added support for LaTeX autocompletion in Jupyter Notebooks (.ipynb files)
+
+## Changed
+
+- Better detection of LaTeX formulas in Markdown cells
+- Improved activation logic to ensure the extension loads for notebook files
+
 ## v0.0.2
 
 ### Changed